StudentHandler update
학생들의 성적을 입력하고, 성적 총점, 평균, 전체 데이터를 볼 수 있는 프로그램 작성

+++ java_project/Java_Study/src/work0529_3/Student.java
@@ -17,7 +17,7 @@ public class Student {
 		return kor + eng + math ;
 	}
 	public float avgScore() {
-		return (float)((kor+eng+math)/3);
+		return (kor+eng+math)/3.0f;
 	}
 	public void showAllData() {
 		System.out.println("이  름: "+ name);
